Introduce TaggedAbstractTensor
This generalizes `AbstractTensor` to the templated struct `AbstractTensorWithInfo<Info>`.

Related to #2913.
